Unit test moved to new class in Bug folder
